moved filter_2d header to details/
